Refactor tests into module
* Attendee refactor according to pep8
* Preliminary onsite form work
* Fix date hardcoded (closes #81)
