+++ registration/payments.py
@@ -1,3 +1,4 @@
+import json
 import logging
 import uuid
