+++ registration/views/common.py
@@ -1,3 +1,4 @@
+import json
 import logging
 from datetime import datetime
