+++ registration/views/onsite.py
@@ -1,3 +1,4 @@
+import json
 from datetime import datetime
 
 from django.shortcuts import render
