+++ registration/views/upgrade.py
@@ -1,3 +1,5 @@
+import json
+
 from common import (
     clear_session,
     get_client_ip,
